Ensure int is passed as microseconds parameter.

+++ peewee.py
@@ -4076,7 +4076,7 @@ class TimestampField(IntegerField):
             elif self.resolution > 1:
                 ticks_to_microsecond = 1000000 // self.resolution
                 value, ticks = divmod(value, self.resolution)
-                microseconds = ticks * ticks_to_microsecond
+                microseconds = int(ticks * ticks_to_microsecond)
                 return self._conv(value).replace(microsecond=microseconds)
             else:
                 return self._conv(value)
